Move client object stores services to src/api
